ldif/scripts/depth_requirements.txt
absl-py==0.9.0
astor==0.8.0
certifi==2020.4.5.2
cycler==0.11.0
decorator==4.4.2
freetype-py==2.2.0
gast==0.2.2
glcontext==2.3.5
google-pasta==0.2.0
grpcio==1.27.2
h5py==2.10.0
imageio==2.8.0
joblib==0.16.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.3.1
Markdown==3.1.1
matplotlib==3.3.4
mkl-fft==1.0.15
mkl-random==1.1.1
mkl-service==2.3.0
moderngl==5.6.4
networkx==2.4
numpy==1.18.1
opt-einsum==3.1.0
pandas==1.0.4
Pillow==7.1.2
protobuf==3.12.3
pyglet==1.5.23
PyOpenGL==3.1.0
pyparsing==3.0.7
pyrender==0.1.45
python-dateutil==2.8.2
pytz==2020.1
PyWavelets==1.1.1
scikit-image==0.17.2
scipy==1.4.1
six==1.15.0
tabulate==0.8.7
tensorboard==1.15.0
tensorflow==1.15.0
tensorflow-estimator==1.15.1
tensorflow-hub==0.8.0
termcolor==1.1.0
tifffile==2020.6.3
tqdm==4.46.1
trimesh==3.6.43
webencodings==0.5.1
Werkzeug==0.16.1
wrapt==1.12.1
